Export goals, use machine prefix
Resort to exporting goals, now machineGoals, but still only use it
directly once.

Prefer to prefix things specific the this sdm with "machine" rather
than "demo" or no prefix.  Using "demo" creates spurious differences
between SDMs.  Using no prefix confuses core SDM types and those
defined in this machine.

Move push test to its own file.  Rename demo.ts to support.ts.
