+++ lib/machine/push.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright © 2019 Atomist, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+    isMaterialChange,
+    not,
+} from "@atomist/sdm";
+
+/**
+ * Immaterial change push test.
+ */
+export const ImmaterialChange = not(isMaterialChange({
+    extensions: ["java", "html", "json", "yml", "xml", "sh", "kt", "properties"],
+    files: ["Dockerfile"],
+    directories: [".atomist", ".github"],
+}));
